Google Closure mode should use goog.bind rather than inlining a JS utility.
This fixes problems in advanced mode compilation when using the fat arrow at the same time as importing a namespace with alias.
The fat arrow adds a __bind function, but the import compiles to a goog.scope, in which you're not allowed to declare new vars.

See nice goog scope article here:

    http://shifteleven.com/articles/2011/08/25/aliases-with-google-closure